backend/requirements.txt
Django==2.0.6
djangorestframework==3.8.2
pytz==2018.4
wheel==0.24.0
